requirements.txt
linkedin-api==2.0.0
python-dotenv==0.19.0
openai==0.27.8
fastapi==0.104.1
uvicorn==0.24.0
python-jose==3.3.0
pydantic==2.4.2
python-multipart==0.0.6
schedule==1.2.0
requests==2.31.0 